pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["EnzoFerreira <enzoraci21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
python-decouple = "^3.8"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
